backend-python/requirements.txt
pymongo==4.6.2
requests
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
transformers==4.37.2
torch==2.2.0
faiss-cpu
numpy==1.26.4
pydantic==2.5.2
python-multipart==0.0.9